requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
jax==0.4.34
jaxlib==0.4.34
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
rich==13.9.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sounddevice==0.5.1
sqlparse==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
